Added twisting, gradient BG, knob colour animation, blinking, breathing; Changed pulse and shiver to CSS animation
Tidied up a few ideas by using CSS animations more for the various effects - at the very least leads to cleaner test code while I'm working this stuff out, though it's clearly less flexible programmatically if I want to tweak values based on what's happening in the game.

Currently I'm seeing myself as just adding a bunch of basic output modalities to start to get a sense of how it feels. Already wondering whether I'm straying too close to a human-like output form (notably the breathing and blinking). Also wondering if it's just too over the top right now.

A good choice is always to strip back, so I may want to take that step again soon, but I do need to make sure I have a way of making the interaction actually feel like something. It still feels a little frictionless and a little undirected. It would make sense for the interface to have desires that can be communicated, rather than be a black box?

In which case literal instructions? "Slide me between 3 and 5"? That's very sexual immediately. Things to think about. Really need to write a journal entry about this at this point.