Moved to Windows 95 styling, added dialog generation from Work game to test, moved to 10 settings instead of 100, added local versions of libraries
? There's a distinctly different feeling involved with the win95
aesthetic. I'm not 100% sure that it makes total sense to me yet, but
it's approaching something that's looking more like it has a personality
and some kind of grasp on a reality. It's not so clear that this
aesthetic 'make sense' for this application, but even just to position
it as a proper sequel/co-existent of It is as if you were doing work
feels like a valuable addition.

? In moving back to the "only 10 settings" version of the slider the
experience is of course much more janky and unsmooth... but I'm
beginning to wonder whether there's something nice about that. It brings
the computer back in a way, less evocative of a biological pleasuring
and more of a computer pleasuring with ten unit settings. Makes more
'sense' perhaps in terms of this balancing act between wanting it to be
actually sexy and wanting it to be actually computational feeling. Hard
line to walk.

? Confirmed that when a dialog pops up it doesn't steal focus from the
slider, which means it's possible to bring up dialogs during play
without ruining the user experience. Pondering the idea of dialogs that
aren't dismissable and just represent an output for the system to
indicate how things are going. They can disappear on their own over time
and the player could move them out of the way. Have to consider, though,
how this would play out in mobile? The dialogs may be too dominant in
that scenario.

? Continue to lose serious time in CSS styling woes with these kind of
overcomplicated libraries to emulate uncomplicated UI setups. There's
probably some delicious comical irony in there, but it doesn't register
as fun while I'm doing it.