Created 'app version' with the UI inside a window, styled instructions/feedback, added progress bar
# App version

As per the design journal, I've been feeling like the slider on its own in the centre of the screen looks too isolated and too close to a potential 'slickness' that seems to reflect a more contemporary style and thus to evoke associations/connotations/expectations I don't want. By moving the slider and other UI inside a typical win95 style window, I think that assumption gets broken pretty effectively. I actually quite like the observation that the title in the menu bar gets abbreviated with an ellipsis on mobile, that has a realness to it that I appreciate.

# Styled instructions/feedback

In keeping with this aesthetic change, I gave the instructions a bevelled border in the same way as the slider, mimicking the win95 approach. Made the background of the text field grey because I feel like that indicates text that is output from the computer, that could conceivably change, but that cannot be edited? Not sure where that assumption is coming from, particularly the idea that it can indicate something dynamic. Log files maybe?

# Progress bar

Added a progress bar officially and again styled it as per win95 (they have a white background, garishly). The progress bar is kind of a nice symbol of the simplification of this game, moving away from ambiguity and the sexiness of not knowing to certainty and the sexiness of knowing. Makes my life easier and it's good to have that symbolised right there in the UI in case I forget!