Added a minimum number of frames to hold selection for it to count, fiddled with target formatting
# Selection timer

Trying to avoid the situation where someone can just blinding slide the slider up and down and not care about what it wants. By setting a minimum number of frames you have to hold the slider in position, you can't do that any more. By making that number of frames quite low, you can maintain a good fluidity of motion if you want to, and this allows for the idea of being too fast as well as too slow - it feels less plodding. Especially if I increase the number of desired strokes to a higher range so you can get into a good rhythm for a while before it requests something new. Sexy?

# Target formatting

Had a conversation with Jonathan over coffee about how to show the target. It's kind of painful, but you also kind of accept whatever it is. I added a border to the selected one which kind of feels better somehow, less arbitrary? Could even be that it doesn't need anything more than an outline, not even a background colour. Hard to say, but I'm fairly happy with it for now.