Did initial work to make it cross-platform, added modal 'tell me you love me' dialog
# Cross platform

Basically just shifted it over to a set pixel width instead of using viewport. It's hacky but it does look pretty acceptable across platforms now and it's exact. I'll have to test it on a real phone, and for some reason it's not centring on an iPhone SE (in emulation), but largely it seems pretty effective and it would be a huge relief not to have to fuck around with this too much.

# Modal talk dialog

Added a basic dialog box that pops up over the interface and interrupts everything (including sending a fake mouseup to the slider so you can't keep or resume sliding) and asks you, for now, to write "I love you", though of course I'll want that to be minimally procedural. I like the idea of an interrupt like this to break the flow of things, and it's a key opportunity to have the player express themselves in a different way. It doesn't contribute to arousal but is rather a way to have a different form of input that commits you to the experience more...