Added sound effects, music and a embryonic title screen
# Sound

Took the first step of inserting a bunch of Windows 2000 sounds into the game in 'appropriate' situations. Some of them match the _Work_ soundscape, such as the dialog-opening sound and the correct/incorrect button sounds. But there are new ones like a click for acknowledging the slider when to the right location, a notification sound for when the range shifts and/or you make progress on the progress bar, a startup sound that plays while a title card fades in.

# Music

Created some smooth-ish jazz on [Wolfram Tones](http://tones.wolfram.com/generate/) and have that playing after the title when the application itself opens. It doesn't sound _quite_ right yet, but I think it makes a decent amount of sense. I guess you could have a button/check-box that enableds or disables 'mood music'? Might be pretty funny. Default it to off. Yeah that makes better sense. Here we are, designing in a commit message, just like MDMA taught us. Getting the music to loop clearly required installing [howler.js](https://howlerjs.com/) which seems to do a really nice job of it, god bless other people.

# Title screen

Just an idea for now, but as a consequence of having a startup sound you need an actual startup. Earlier I'd thought about maybe presenting a desktop icon for the application the user would start up (and still could need to do this pending issues around sound triggering on iOS for example?), but for now I'm going with the way shitty Windows programs load with that awkward application card that appears in the middle of your screen while the application loads. So I've got a kind of simulated version of that for now that fades in to the startup sound, then goes away when the app window pops up.