Added icon that launches the application, improved slider positioning and dialog positioning
# Icon

The game now starts as an icon on an empty 'desktop'. The current icon is taken from some semi-random set of Windows 95 (ish?) icons, it's an edited version of a human hand holding up a printer. I removed the paper coming out of the printer so now it looks like (to me) a human hand holding a computer, which is kind of a nice 'loving' image perhaps? I'm not 100% happy with the aspect ratio of the image, because it's a bit flat and doesn't quite read as much like an application icon as I want. I can fiddle with that.

The icon-launching version did turn out to be necessary because iOS won't let you play sounds without some kind of input event. So now clicking the icon is the thing that starts everything rolling and from there it should just work.

# Positioning

Nothing so interesting, just making things work better.