Added more controlled stroke ranges and strokes required, stroke timer, rearranged code, added title splash
# Controlled strokes

Now have stroke ranges and numbers set up with arrays so that it's easier to tune to specific stroke formats and numbers, it works well I think and makes more sense in terms of being able to control the experience the player's having - there are particular kinds of strokes that are sexier, I think, than others (e.g. 0 to 10 is hotter than 1 to 7, say). I mean, this is getting pretty weird, but I think it's true. Likewise, the number of strokes required is a finesse thing - you particularly don't want too few I think, though perhaps occasionally it's good.

# Stroke timer

Not doing anything with it yet, but the code now counts frames between selections so you know how long a stroke took. I was impressed by how it felt like something even to do that. And actually this makes me wonder whether actually displaying the current average stroke time or something makes some sense? Will think about it. Could be part of the feedback? Like "slow down, you're going at X-strokes-per-second"? This also points me at another problem which is how likely the player is to pay attention to the speed feedback. Should that be more front and centre with the slider? I feel like I'm gradually obviating the instruction panel though? Need to think about that more...

# Title splash

Wanted to have one of those retro 'card style' splash screens while the application loads. Ended up using Microsoft Windows 6.0 as a reference point, which has mostly text and then a clip-arty drawing of a pen (e.g. a kind of indexical sign?). Created something roughly like that but with a picture of a human heart (pixellated). Originally used a really amazing, bloody pig's heart an artist had taken a photo of, which looked very good, but for copyright purposes went across to a royalty free image which works well too. Definitely adds a lot to the idea of this as 'real' software, so I think that's a big win. It's fun.

Note that in using the image of a heart as an organ I'm sort of gesturing toward this idea that the software is machine-made in some way and that they've made a slight error in picking the wrong kind of heart. It's not clear to me how far down that road I want to think (like 'what would the AI do/design?'), but I find it amusing, and it signals something 'weird' is going on from the beginning. I wonder now whether the icon for the application should be the/a heart too?