Added tracking of stroke speed, changed to a single "messages" panel
# Stroke speed

Added the idea of fast/slow/indifferent and timing how long each stroke is. Track number of 'bad strokes'. Set up to be able to give feedback ot the player based on all of this.

# Message panel

Started to feel like it was redundant having the two panels. And 'messages' is a really fun name for an output.