Added breathing sounds
# Breathing

I added breathing sounds because breathing is a pretty important part of making something feel organic, and of course because breathing is a big part of representing exertion and, in this case, arousal. It's a way to make the game sexier outside the obvious confines of the user-interface which by its nature needs to be pretty blank. I am, however, still using a Windows sound as the breathing - it's called ir_begin (presumaby played when an ifra-red thing is about to transmit, all very erotic in itself). By playing with the pitch and pacing and volume of the sound I think I'm able to get something that sounds fairly close to breathing, which is neat. By speeding it up and pitching it up etc., there's a definite sense (I think) of increased arousal from the interface... it takes it somewhere much more intimate and gives you some actual feedback on how things are going!