Added text input request with random requests and hints, added positive feedback phrases (but not implemented yet)
# Text input

Created an array of text input requests designed to help the player feel more involved by 'expressing' themselves in words. "I want you", "I need you" and so on. Did some of the obvious regular expressions stuff to give the player _some_ leeway with how they type it in, but also like the idea that they have to type _exactly_ the right response. If you get it wrong twice the system will just tell you what to time, which is also funny, like someone guiding your hand - "ugh, just do it like this!"

# Positive feedback

Added a bunch of phrases that can be popped up at random intervals to emphasise the sensuality of the experience for the computer side (or at least its feigned engagement at that level). Have to continue to think about the language here as right now it's _really_ suggestive, like "keep sliding me"... need to think about whether I want that or whether it should be a little less obviously kind of human. Should it use metaphors and language from an operating system/piece of software? It's a question.