Added feedback messages, feedback dialogs, arousal entropy
# Feedback messages and dialogs

The game now gives feedback for poor strokes (too fast, too slow) and good strokes (on target). Feedback can be either an update in the message field (including a flash so you actually check it), or a dialog box if it's particularly important (if you have multiple bad strokes or if you're doing well over time). It feels like it makes a big difference to the experience and the flashing message box actually works really well for foregrounding that UI element and helping the player pay attention to it. Who knew that a tried and tested form of UI design would be effective? Eh? Not me.

I have concerns/thoughts about tuning the language, and there's of course tuning in terms of the frequency of the dialog based feedback in particular since it's pretty disruptive. Though they at least don't interact with the stroke timer, so you don't get penalised in that way (and notably it doesn't affect entropy).

# Arousal entropy

After thinking I wouldn't have it in the game, I've added the idea that arousal diminishes over time if you just leave the system alone. There's a threshold and then it starts to decline, based on the current stroke time. So if you just stroke and then don't do anything, then eventually arousal drops toward zero. But there's room for a little breather at least. Note that this conflicts with an idea I had that there would be an instruction to just leave the slider completely alone as an erotic act, but I think maybe that's too complex at this point, too 'hard' for the user to do? I'll think about it.