Added the orgasm
# Consider the orgasm

Hoo boy, that was a flurry of programming. Integrated pretty much what I said I would in the todos so that the game now actually has an ending. So when you reach an arousal level of 1.0 you get a kind of 'spasming' of the various UI elements that 'makes sense' for - sounds, the slider setting, the slider target, the radio buttons, and the messages field. I think there's some work to be done in terms of tuning it, particularly the idea of the cooldown, which right now feels not ramped quite right - I think we need to dwell in the end of the cooldown longer, those final twitches. But overall the orgasm state is... honestly it's pretty good.

# Consider the post-orgasm

Right now once the orgasm is done the whole app shuts down, returning to an earlier idea I had that it would be kind of 'selfish' about things and just leave on completion. I mostly did this because... it's easier right now and it feels like a complete narrative moment? So it shuts down and then there's a "Thanks for playing!" dialog that pops up, and when you close that it's the shutdown sound and that's that, no ability to restart the game or anything.

I'm wondering about

- Do I implement the version where you can make a date to meet again?
- Do I implement the idea of the UI having a cooldown between sessions so you can't just reload it and go again? (Is this interesting in this context, or just me wanting to be Robert Yang?)

# Language

The language needs a buuuuunch of work mon frère. I know that and you know that.