Added readme, readme icon, new app icon
# Readme

As with the Work game, I wanted to have some kind of diegetic text that explains the nature of the application and situated the speculative play aspect of things for anyone who might be interested. Again it's a kind of readme.txt you can check out before playing, totally optional. I mirrored the style of the Work text and I think the argument made is pretty reasonable (I mean makes sense, obviously). It was a real challenge getting a text that would be short enough to fit on a mobile screen and the iPhone 5/SE still elludes me (for the entire UI acutally, not just the readme).

# Icons

I used the standard windows "text file" icon for the readme, unsurprisingly. Once I had that in, the app icon looked a bit weird and not so nice, just because I'd repainted it rather than using a pure icon. As such I went back to original icons and found a stand-alone hand and a computer with what kind of looks like fireworks on its screen (maybe for a screensaver or something?). I combined those in Pixen to end up with a hand kind of caressing the computer which errupts in fireworks? At least that's what I see.