Built a horizontal slider prototype
# But why?

I'm in conversation with Jim and Mary (my parents) about the gendered nature of the vertical slider. Had the important experience of Mary pointing out that sliding the vertical slider has different connotations for her as a woman. I'm trying to walk this line of the interface evoking the biological kind of sex and drawing on the eroticism of that, but without wanting to slide into the 'masturbating a male interface' chasm because that will really alter the experience for a lot of people and not in a way that I think is the right direction for the overall idea behind this project. Specifically, if I want it to feel like 'pleasuring a computer because pleasuring other people (or things?) is a pleasure' then it needs particular kinds of inflections. If it feels like a chore, or a gross patriarchical or gender normative thing, then it's going to lose that sense.

So I've made the slider horizontal as a test to run it by people. It maintains the important (I think) quality of still being a real-time and continuous form of input (unlike other UI elements) which I find important or at least interesting for the sexual nature of things. But by not being vertical it's at least not immediately like an erect penis? I mean, who ever thought I'd be writing about things like this in a git repository commit message? Not me. But it's also kiiiind of amazing I haven't explicitly addressed this much until the 'end' of the project. We'll see whether this issue ends up requiring more radical redesigns, but it's becoming clear that I suspect can't just go ahead with the vertical... argh.