Created a separate data file for text, rewrote all texts, adjusted stroke difficulty, made multiple games possible
# Rewriting

Pursuant to the large amount of soul-searching about language alongside good feedback from a number of sources (notably Jim, Mary, and Jess), I tried my hand at a serious rewrite of the language in the game to set a more 'accurate' tone to the idea behind the software. I think it's significantly better, but it needs to be tested. (I won't lie, though, I'm feeling pretty fatigued by this at this point, so can't guarantee I've genuinely done a good job.)

# Stroke difficulty

Jim and Mary both had trouble with the fast strokes, so I made them easier - it's not really about being 'good at it', just about distinguishing fast and slow.

# Multiple games

The icons don't vanish from the desktop and clicking the app icon resets the game so you can play again right away if you want. That kind of accessibility makes a lot more sense to me in the better-defined context of the game. It _is_ in service to the human (even as they service it).