Made about text smaller to fit vertically (on my phone), fixed typo, trying a 'no scroll on mobile' approach (position fixed for body and html)
# All about mobile

I really want the game to work well enough on mobile that somebody who chooses to play it that way doesn't view it as utterly broken. I already know that on the iPhone SE (and lower I guess) it doesn't display well because the width of the display is 20px to narrow, leading to overflow. The game is still playable in that scenario, though (no vital UI is blocked), so I'm pretty tempted to not give a shit.

However it would be nice to eliminate the frustrations of the way iOS hides and shows the browser chrome depending on scrolling, so I'm trying to kill scrolling with A Tip From The Internet. Worked in Chrome emulator, but that means exactly jack shit of course. Pushing mostly to test it.