Code rewrite to move selection results into a timer set off in slide to attempt to get more iOS sounds working
# Why?

iOS only allows sounds to play when they're triggered by a user event. But I need sounds to play on a delay from user input (e.g. selection is acknolwedged a little bit after you select to create the effect of needing to be purposeful). I get the impression that iOS will play a sound triggered by a timeout that's set inside an event handler. It also results in kind of cleaner code I thiiiiink? Anyway, I did this to test my hypothesis and to thus get the feedback 'click' on iOS.