Wrote a better game over, moved and rewrote a bunch of code to try to force iOS to play my sounds (specifically the orgasm cacophony), added shake effect for a trembling pre-orgasm moment that quite pleasing
# Game over message

Game over is now an "arousal overflow" which I find pleasing. It's an ending dialog based around a screenshot I found of a Windows-based buffer overrun error. It works nicely as a deadpan ending, with the app itself still orgasming all over the place in the background. When you dismiss the message, the whole thing shuts down. I actually only did it this way in order to be able to turn the sounds off! But it works well.

# Pre-orgasm

Same thing as above. iOS sound restrictions require me to induce a click from the user so that I can turn sounds/on off. Notably if I want to set the orgasm sounds in motion, there needs to be a specific user input (a click), so now there's a trembling pre-orgasm dialog (may need better language) that sets it off when you say 'ok'. It feels a little bit teasing/s&m in a way? Maybe it should say "May I?" Huh.