Edited about text, removed positive feedback and input dialog boxes
# About text

Just some tweaks to try to really nail the language. Think it's pretty much there now.

# Dialog boxes

Decided after some testing with J+M that they're too intrusive when they show up while you're being successful. The positive message is just a distraction and breaks the flow and you already know you're doing well from the other messages. The text-input dialog is a funny thing and worth exploring, but in this game it feels like it's too much or pulling to hard in a different direction. I'll use it another time.